Fix dbconvert.

+++ src/funmath.c
@@ -659,13 +659,13 @@ void fun_div ( char *buff, char **bufc, dbref player, dbref caller, dbref cause,
 
 void fun_floordiv ( char *buff, char **bufc, dbref player, dbref caller, dbref cause, char *fargs[], int nfargs, char *cargs[], int ncargs )
 {
-    int top, bot, res;
+    long top, bot, res;
     /*
      * The C / operator is only fully specified for non-negative
      * operands, so we try not to give it negative operands here
      */
-    top = ( int ) strtol ( fargs[0], ( char ** ) NULL, 10 );
-    bot = ( int ) strtol ( fargs[1], ( char ** ) NULL, 10 );
+    top = strtol ( fargs[0], ( char ** ) NULL, 10 );
+    bot = strtol ( fargs[1], ( char ** ) NULL, 10 );
 
     if ( bot == 0 ) {
         safe_str ( "#-1 DIVIDE BY ZERO", buff, bufc );
@@ -711,13 +711,13 @@ void fun_fdiv ( char *buff, char **bufc, dbref player, dbref caller, dbref cause
 
 void fun_modulo ( char *buff, char **bufc, dbref player, dbref caller, dbref cause, char *fargs[], int nfargs, char *cargs[], int ncargs )
 {
-    int top, bot;
+    long top, bot;
     /*
      * The C % operator is only fully specified for non-negative
      * operands, so we try not to give it negative operands here
      */
-    top = ( int ) strtol ( fargs[0], ( char ** ) NULL, 10 );
-    bot = ( int ) strtol ( fargs[1], ( char ** ) NULL, 10 );
+    top = strtol ( fargs[0], ( char ** ) NULL, 10 );
+    bot = strtol ( fargs[1], ( char ** ) NULL, 10 );
 
     if ( bot == 0 ) {
         bot = 1;
@@ -742,13 +742,13 @@ void fun_modulo ( char *buff, char **bufc, dbref player, dbref caller, dbref cau
 
 void fun_remainder ( char *buff, char **bufc, dbref player, dbref caller, dbref cause, char *fargs[], int nfargs, char *cargs[], int ncargs )
 {
-    int top, bot;
+    long top, bot;
     /*
      * The C % operator is only fully specified for non-negative
      * operands, so we try not to give it negative operands here
      */
-    top = ( int ) strtol ( fargs[0], ( char ** ) NULL, 10 );
-    bot = ( int ) strtol ( fargs[1], ( char ** ) NULL, 10 );
+    top = strtol ( fargs[0], ( char ** ) NULL, 10 );
+    bot = strtol ( fargs[1], ( char ** ) NULL, 10 );
 
     if ( bot == 0 ) {
         bot = 1;
